Improve `unregister` API and internal helpers
- Modified `unregister` API to receive a `Service.Type` parameter
instead of a `Service`, to allow unregistering all types of services.
The previous implementation didn’t make much sense, given that receiving
a `Service` instance meant any lazy service with that type (and name)
would have to already be initialised. This was somewhat silly and also
not very user friendly, because it required the user to have an
instance to unregister it, when we only need it’s type (and name).
Also avoided initialising the lazy service when the name matched to
validate its type, making the type check using
`LazyInit<Service>` instead.
- Simplified some helper methods’ names.
- Created new `lazyServiceTypeMismatch` `ServiceLocatorError` type.
- Renamed `LazyServiceClosure` type alias to `LazyInit`
- Changed `lazyServiceInits`’s type to `[String : Any]` to avoid seg
faulting the compiler when trying to cast from `LazyInit<Any>` to
`LazyInit<Service>` on validations.