Add `CoreDataStack` 💾
- Created `CoreDataEntity` protocol to contain some utilities for common
operations which aren’t yet available natively on iOS 8 and 9.
- Created `CoreDataStack` protocol to represent a Core Data Stack,
together with a set of extensions to create the required
components of a Core Data stack, as well as methods for basic CRUD
operations on a generic `CoreDataEntity`.
- Created `NestedContextCoreDataStack` for a nested context Core Data
stack implementation, a `MainQueueNestedCoreDataStack` subclass and a
`PrivateQueueNestedCoreDataStack` type alias.
- Created `DualContextCoreDataStack` for a nested context Core Data
stack implementation and a `MainQueueDualCoreDataStack` type alias.
- Created `ManagedObjectReflectable` protocol to enable the reflection
of one type to another.
- Created a `CoreDataStack+ManagedObjectReflectable` protocol extension
to specialise CRUD operations for any `ManagedObjectReflectable` and
abstract the reflection logic between both Entities automatically ✨.
- Created Unit Tests for `CoreDataStack` setup and operation methods.
- Created `MockErrorManagedObjectContext` to mock errors on a context.
- Created extension methods on `NSManagedObjectContext`
- Created extension method on `NSPersistentStoreCoordinator`
- Created extension init on `NSPersistentStoreDescription`
`performThrowing` and `performThrowingAndWait` which are throwing
versions of `perform` and `performAndWait` that also return a generic
value `T` 💪