Improve `URLSessionNetworkStack` (#51)
In order to define `self` as the session's delegate while preserving
dependency injection, the session must be injected via property. This
is because the session's delegate is only defined on its `init`. 🤷‍♂️

The session's delegate could be set to `self` using a lazy var (since
`self` is already defined), but then the session couldn't be injected
for unit testing.

Added a new optional property to the `Network.Configuration` to perform
authentication challenge validations. This is a closure property which
receives the `URLAuthenticationChallenge` from the session’s
`URLSessionDelegate`’s
`urlSession(:didReceive challenge:completionhandler:)` method and
invokes the supplied `completionHandler`.

Until we have a proper server trust validator, this closure can be used
to perform any validations like allowing servers using expired and
self-signed certificates.