Enhance `View` create `(Reusable)ViewModelView`
The goal of this family of protocols is to enforce a common base and
lifecycle for all views in a project.

- Modified `View` protocol to have functions `setUpSubviews()` and
`setUpConstraints()`instead of simply `setupLayout`.
- Created `ViewModelView` and `ReusableViewModelView` protocols which
extend `View` and `ReusableView`, respectively. They should
receive a `ViewModel` on `init` and have a `setUpBindings` method. This
enables MVVM views to conform to `ViewModelView` and have a ViewModel
injected (otherwise they couldn’t use `View` with constructor-based
dependency injection).
- Updated CollectionView TableView cells and header/footer views
- Updated folder/project structure and renamed `Protocols` folder to
`View`.
- Added some UT’s for the CollectionView and TableView cells and
headers/footers.
- Improved `UICollectionView` and `UITableView` `ReusableView`
extension methods error messages, and made calls fail with `fatalError`
instead of `assertionFailure`.