Add custom API Error parsing to `NetworkResource`
Whenever a request fails with a non successful status code, the payload
of the request was being ignored, and no attempt to parse error
information was made.

By adding an error parsing closure `(Data) -> E?` to the
`NetworkRequest`, the `NetworkStack` can then be try to parse error
information from the returned data, if any.