Improve JSON parser API's and add optional parse
On some scenarios, using type inference on the result type alone is
cumbersome, e.g. makes chaining parse calls impossible, and using it on
`if` or `guard` statements is also ugly.

Another common used approach (also used in Swift 4's new `Codable`) is
to receive the type as a parameter of the API itself. This makes the
API more flexible and solves the above issues.

To address these limitations, two new API's were added that work over
specified types instead of inferred ones.

Additionally, some models have optional attributes, which can appear or
not in the JSON. This was not properly addressed in the current
implementation, so a new API `parseOptionalAttribute` was created to
specifically handle these cases. When the attribute key doesn't exist,
an attempt is made to parse a custom API error, and if it doesn't exist,
 `nil` is returned. If, on the other side the key does exist,
validation is made like on `parseAttribute`.