Add completion closure to `RouteHandler` 📥 (#75)
For some use cases, it is handy to know when a route has effectively
been handled by its `RouteHandler`, instead of losing track of the
execution flow and simply delegating the handling logic.

To address this, a new optional completion closure parameter of type
`((T) -> Void)?` was added to `RouteHandler`'s and `TreeRouter`'s
`handle` methods. Consequently, `RouteHandler` gained a new associated
type `T`, and `TreeRouter` a new generic constraint.

Using a generic type seemed like a nice balance because not only is
completion signaled to the `route` method caller, but the flexibility
to return a custom payload is also provided. This allows the Routing
infrastructure to be used in even more scenarios.

Fixed some synchronization issues on Logging tests and `StoreTestCase`
 which were breaking Travis builds.