Recreate request on network fetch retry 👯
Whenever a request had to be retried after being re-authenticated, the
original `URLRequest` was simply being reused, instead of being
recreated.

This prevented reacting to the re-authentication event and adjusting the
generated request for that particular resource.

By preserving the original `NetworkResource` throughout the calls, we
are able to call `toRequest()` every time the request is (re)made.