Send original request to `shouldRetry` ♻️
When a request fails and `NetworkAuthenticator`'s `shouldRetry` is
invoked, information about the original request can be very useful to
pass on, since it enables better decisions and optimizations.

A simple example is when multiple requests fail due to an expired
token. In this scenario, only the first should trigger a token renewal,
and the remaining ones should use the token fetched by the first
request. Using the original request data, we can accurately compare the
tokens of each request (and/or additional metadata) and check whether
the token renewal is in fact required or not.