Rename `NetworkAuthenticator` API
Since retries might not be tied solely to authentication in the future,
the `shouldRetry` API doesn't make much sense, and was renamed to
`isAuthenticationInvalid`, to represent its function more clearly.

In the future, when we implement a proper retrying mechanism we can
simply use the authenticator's check as an input to count/allow retries.