Improve and rename `shouldRetry` ♻️ (#100)
When a request fails and `NetworkAuthenticator`'s `shouldRetry` is
invoked, information about the original request can be very useful to
pass on, since it enables better decisions and optimizations.

A simple example is when multiple requests fail due to an expired
token. In this scenario, only the first should trigger a token renewal,
and the remaining ones should use the token fetched by the first
request. Using the original request data, we can accurately compare the
tokens of each request (and/or additional metadata) and check whether
the token renewal is in fact required or not.

Additionally, since retries might not be tied solely to authentication in 
the future, the `shouldRetry` API doesn't make much sense, and was 
renamed to `isAuthenticationInvalid`, to represent its function more 
clearly.

In the future, when we implement a proper retrying mechanism we can
simply use the authenticator's check as an input to count/allow retries.