Improve `Router` and `TreeRouter` API’s
- Allow multiple `RouteHandler`'s per `TreeRouter`

Since `TreeRouter` was generic on `RouteHandler` and `T` (because of the
completion closure having a `T`), only a single `RouteHandler`
implementation could be used.

To enable multiple handlers on a single `TreeRouter` (simplifying it to
`TreeRouter<T>`, a new type erasing `AnyRouteHandler<T>` was created
to wrap any handler implementation using a `T`.

- Improved `TreeRouter` API

  + Added `@discardableResult` to `unregister`
  + Changed `completion` on `route` API to `handleCompletion`, to be
clearer as to what it means (i.e. it's only run when a match is
found and the handler executes)

- Removed `register` and `unregister` from `Router` API