Improve `DiskMemoryPersistenceStack` 📼
- Revised `DiskMemoryPersistenceStack`'s concurrency to allow
injecting custom `DispatchQueue`s for reading and writing, with the
`readQueue` being optional so that when set to `nil` reads from disk
are made *synchronously*, avoiding a context switch (queue jump).

- Updated `PersistenceStack` to use `Result` based completion
handlers, and to have an associated Error type (instead of forcing
`Persistence.Error` to be used).

- Removed `Persistence.Error`, since now cache misses are
represented as returning a `nil` value (instead of
`Error.noObjectForKey`, and we avoid an error nesting that didn't
bring much value (using `Error.other`).

- Made initial used disk space calculation synchronous.

- Made `usedDiskSize` thread safe.

- Nested `DiskMemoryPersistenceStack` inside `Persistence` namespace.

- Did some tweaks on `NetworkPersistableStore` to improve error
handling and fetches.