Refresh cache on `.persistenceThenNetwork` cache hits
When using the `.persistenceThenNetwork` strategy on
`NetworkPersistableStore` and the data existed on the cache (cache hit),
we were never refreshing the cache by performing a network fetch in the
background. This ultimately resulted in the cached data becoming stale
until it was evicted (which in the case of disk cache could be a long
time).