Fix retry after delay conversion (#160)
On 32 bit devices, the conversion from `ResourceRetry.Delay`
(`TimeInterval`) to nanoseconds was greater than `Int32.max` if the
value was greater than `Int32.max/NSEC_PER_SEC` (2.147483647), which
resulted in a runtime crash.

After closer inspection, we realized that we can simply send in the
seconds directly to `asyncAfter`, and simply add them to `.now()`. 😅

Additionally, a slight optimization was made so that on negative or 0
delay we simply retry immediately after (we also don't add possibly
negative values to `totalRetriedDelay`.