Make `NetworkStack` conditionally a `NetworkStore`
Created a protocol extension on `NetworkStore` when `Self: NetworkStack`
and the Error is `NetworkPersistableStoreError`. This enables using a
`NetworkStack` implementation (e.g. `Network.URLSessionNetworkStack`)
as a `NetworkStore` directly (more like a `NetworkPersistableStore`),
when one doesn't want/need persistence. This avoids having the overhead
of using a `NetworkPersistableStore` with a dummy persistence stack.

Cleaned up `Network.URLSessionNetworkStack` internal API's to not have
to pass `apiErrorParser` around, since it's on the resource.

Added `NetworkStoreTestCase`.