Improve cancellation handling in `NetworkStore`'s (#173)
## Motivation

With the recent resource retry capabilities, the
`URLSessionNetworkStack` fetches now fail with the `.retry` error by
default if a retry policy returns a `.noRetry` action.

This means that if a resource fails with a `URLError.cancelled` error
and a policy handles it returning `.noRetry`, it is not the "vanilla"
`.url(URLErrror.cancelled)` it was before, and instead is a
`.noRetry(_, _, .custom(URLError.cancelled))`.

This difference meant that our `NetworkStore`'s didn't handle these
errors as `.cancelled` as they should.

To make this more robust (i.e. not dependent on `URLError.cancelled`),
the error handling now looks at the `cancellable.isCancelled` flag to
determine if the fetch was in fact cancelled or not. This has the extra
benefit of eliminating the "false positive" cases of fetches that fail
with `URLError.cancelled` because of other reasons (e.g. authentication
challenges failures).

## Changes

- Updated `NetworkStore`'s default `URLSessionNetworkStack` extension
and `NetworkPersistableStore` to return `.cancelled` only when the
network fetch fails and the cancelable has been cancelled.

- Added a `+=` operator on `CancelableBag` to wrap `add`.

- Updated some usages of `CancelableBag.add` to use `+=`

- Improved `MockNetworkStack` to only run the fetch via a new
`runMockFetch` API, to solve timing issues when using the cancelable.