Propagate errors on .cancelled network store errors (#174)
With the latest changes, fetches return a `.cancelled` error if their
`Cancelable` was cancelled and the fetch had failed with an error.
However, this "original" error was not propagated to the outside, which
on some cases might be handy since now more fetches will be marked as
cancelled.

Updating the `NetworkPersistableStoreError.cancelled` error case to
have a new `Swift.Error?` associated value will allow us to propagate
these errors upstream when they occur. The `nil` case is saved for when
no explicit error has occurred leading to the cancellation.