Make `NetworkResource` request creation async 📬 (#179)
Until now, `Resource` fetch authentication was made via a single
`NetworkAuthenticator` injected on a `URLSessionNetworkStack`, which
was used to enrich the requests of the Resources being scheduled on the
stack. This had some limitations:

1. Either all `Resource` fetches were authenticated, or none (depending
on whether `authenticator` was set or not on the stack).

2. Since the `authenticator` was shared when set, supporting different
authentication schemes (e.g. different API's) or unauthenticated
fetches on the same stack required creating a complex
`NetworkAuthenticator` (containing / proxying all logic), which was
obviously not ideal.

To solve the above problems and give more flexibility to our stack, the
`NetworkResource` keeps being responsible for creating its requests,
with the difference that now it does so *asynchronously*. This allows
each `NetworkResource` implementation to create (and enrich) its
requests, only passing them to the stack when they are ready to be
scheduled. As such, more complex request building logic like
authentication can now be done by the Resource itself (which can have
its own `RequestAuthenticator` instance, for example).

## Changes

- Renamed `NetworkAuthenticator` to `RequestAuthenticator` and
`RetryableNetworkAuthenticator` to `RetryableRequestAuthenticator`,
and refined to contain associated types.

- Created specialized `URLRequestAuthenticator` and
`RetryableURLRequestAuthenticator` that work with `URLRequest`, `Data`
and `URLResponse`.

- Updated `NetworkResource` to have a `Request` associated type, and a
new asynchronous `makeRequest()` API.

- Removed `RelativeNetworkResource` and `StaticNetworkResource`.

- Created `HTTPResourceEndpoint` to represent an HTTP endpoint, and
contain default `URLRequest` building logic.

- Created `HTTPNetworkResource` to represent a Resource with a
`HTTPResourceEndpoint`, and contain a default `makeRequest()`
implementation.

- Created `AuthenticatedHTTPNetworkResource` that extends
`HTTPNetworkResource` and has a `NetworkAuthenticator` instance, and
contain a default `makeRequest()` implementation.

- Added UT's for the new types.

- Adjusted existing UT's.

- Cleaned up and unified some `MockResource`s in UT's.

- Added some documentation on the new stuff.