Add `finishFetchesAndInvalidateSession` in stack
`URLSession` objects keep a strong reference to their delegate until
the session is explicitly invalidated.

As such, if the `URLSessionNetworkStack`'s `session` isn't invalidated,
the retain cycle is never broken and the app leaks (until it exits).

This means that this should **always** be made before deinit'ing the
`URLSessionNetworkStack`, to avoid a retain cycle.

Since the stack is effectively the owner of the session, a new
`finishFetchesAndInvalidateSession()` API was added to the stack to
allow performing this operation while serving as documentation.