Rename `DefaultLogger` to `MultiLogger` 📄
In order to be consistent with `PerformanceMetrics` and `Analytics`,
the `Log.DefaultLogger` is renamed to `Log.MultiLogger` since it
applies essentially the same principle of forwarding events to multiple
instances.

Being called "default" could also be misleading for framework users as
it can be overkill for many simpler use cases.