Fix Network playground 🎳
As mentioned in #163, the Network playground has been broken for a while
because of all the changes that have been made on the Network layer.

To address this, the Network playground has been fixed to work with
the latest changes. Additionally, it was updated to perform requests to
GitHub v3 API, which allows users to play with a real API.