Remove iOS 9 support 👋
After some years supporting it, and with iOS 13 right around the
corner, we felt it was time to drop support for iOS 9.

It also allows us to remove some legacy code (and tests), and finally
be consistent in the iOS target version between main target and test
targets.

 ## Changes

- Set deployment target to iOS 10 at project level, and inherit on
all targets.

- Update `s.ios.deployment_target` on .podspec.

- Update Compatibility section in README.md.

- Remove all iOS 9 specific code.

- Revise code to use most up to date primitives.