Revise Network documentation
Like the Network playground, the Network documentation had been
forgotten for a while, and was quite outdated because of all the
changes that have been made on the Network layer.

To address this, the Network documentation has been revised with the
latest changes. Additionally, it was updated to use examples
accessing the GitHub v3 API (like the playground), and also with
sections explaining how authentication and certificate pinning can be
set up.

Closes #163.