Remove TODO and FIXME placeholders (#198)
With the addition of `#warning` and `#error` directives introduced
in Swift 4.2, the `TODO` and `FIXME` placeholders stopped making much
sense to be considered Utils.

Additionally, besides not being used that much (or at all, at least
on our project), they are a pain to test. 😅

That being said, perhaps it’s time to let them go.