Clean up Network errors and type aliases
The `Network.Error` and `Network.Configuration` weren’t particularly
generic, as both were very specialized for working with a
`URLSessionNetworkStack`.

As such, a new `associatedtype Error` was added to the `NetworkStack`
Protocol, and both `Network.Error` and `Network.Configuration` were
nested inside `URLSessionNetworkStack`.

In the same way, the `Network.Value` was modified to also be generic
over the `Response` value, so that it can be used on any other
`NetworkStack`.

The `Network.CompletionClosure` and
`Network.AuthenticationCompletionClosure` type aliases were also
removed and new `typealias`’es were created in each protocol where
they are useful (i.e. `NetworkStack`, `NetworkStore` and
`AuthenticationChallengeHandler`).

The `Network.URLSessionNetworkStack`’s default `NetworkStore`
Conformance was removed, so that it can be used with any other
`NetworkStore` implementation (i.e. with any other `StoreError` type).

To increase flexibility and prepare for upcoming `Decodable` changes,
the `NetworkPersistableStoreError` restrictions on each case’s error
were removed, `.parse` was renamed to `.decode`, and `.other` was
removed.

Additional handling was added to try and parse a custom `APIError`
when the resource’s decode fails inside a `NetworkStore` fetch.