Add `_modify` accessor to `Atomic.value` (#199)
With Swift 5.0 the new `_modify` accessor became available, which
allows a more efficient access to underlying storage when an in place
mutation is made (i.e. via `inout`, like `value += 1`), by `yield`ing
a reference to the storage itself, instead of performing a copy
(`get`) followed by a write (`set`).

This is particularly useful in containers where memory operations can
become expensive (like collections), but also in containers wrapping
expensive operations (like locking/unlocking locks).

Such is the case of `Atomic`, where besides optimizing memory access
to the underlying storage, it can save a lock/unlock when the `.value`
is modified in place (i.e. via `inout`). In this particular case, one
can even claim that it’s more correct, because the whole mutation is
truly made in a single atomic operation instead of two (as is already
the case in the `modify(_:)` API).