Add sizers for collection reusable views 📏
The sizer is used to calculate the size of a view before it is created
and added to the view hierarchy. This is useful in the case of
collection view cells when the layout needs to know the size for a cell
given its view model. It works by wrapping `systemLayoutSizeFitting` on
a dummy view to which the view model is temporarily bound.

This also includes a cache to avoid repeating the same calculations
unnecessarily when a previous value is known to be the same. It uses
a dictionary of size values with a key generated by the view
(conforming to `SizerViewModelView`) for a specific view model.