Fix ReusableView unit tests on iOS 13 /  Xcode 11
Probably due to some underlying changes in UIKit for iOS 13 (Xcode 11),
our tests for `ReusableViewCollectionView` and `ReusableViewTableView`
stopped working because the cells weren't being "drawn" on their
respective container views despite we forcing the `layoutIfNeeded`.

To address this, the respective container VCs were added to a
`UIWindow` as its `rootViewController`, which solved the issue.

## Changes

- Make `ReusableViewCollectionViewTestCase` and
`ReusableViewTableViewTestCase`'s container VC's be added as
`rootViewController` of a dummy `UIWindow`, which is made key and
visible.