Clean up Network errors and type aliases (#202)
The `Network.Error` and `Network.Configuration` weren’t particularly
generic, as both were very specialized for working with a
`URLSessionNetworkStack`.

As such, a new `associatedtype Error` was added to the `NetworkStack`
Protocol, and both `Network.Error` and `Network.Configuration` were
nested inside `URLSessionNetworkStack`.

In the same way, the `Network.Value` was modified to also be generic
over the `Response` value, so that it can be used on any other
`NetworkStack`.

The `Network.CompletionClosure` and
`Network.AuthenticationCompletionClosure` type aliases were also
removed and new `typealias`’es were created in each protocol where
they are useful (i.e. `NetworkStack`, `NetworkStore` and
`AuthenticationChallengeHandler`).

The `Network.URLSessionNetworkStack`’s default `NetworkStore`
Conformance was removed, so that it can be used with any other
`NetworkStore` implementation (i.e. with any other `StoreError` type).

To increase flexibility and prepare for upcoming `Decodable` changes,
the `NetworkPersistableStoreError` restrictions on each case’s error
were removed, `.parse` was renamed to `.decode`, and `.other` was
removed.

Additional handling was added to try and parse a custom `APIError`
when the resource’s decode fails inside a `NetworkStore` fetch.

- Improve `Network.URLSessionNetworkStack.Error`:

  + Unify (again) `.api` with `.http`, as `.api` is essentially a subset
of `.http`, and the benefits of having them separated has proven to not
be that much, especially if we're only interested in the status code.

  + Update `.url` to use a more strongly typed `URLError`, which
apparently is the "official" error type for `URLSession` errors, as
confirmed in:
https://developer.apple.com/documentation/foundation/urlsession/datataskpublisher/failure

- Update Network documentation and playground