Remove `Mappable`, `JSON`, `Parse` and `Serialize` 🚮 (#220)
Since the introduction of `Codable` that `Mappable` became deprecated,
but for one reason or another we didn't remove it. That time has come.

As the `JSON`, `Parse` and `Serialize` "namespaces" are tightly coupled
to the `Mappable` protocol, they have also been removed. The exception
was `UIImage` encoding/decoding to/from `Data`, but that can easily be
done by using native APIs like
`pngData()`/`jpegData(compressionQuality:)` or `UIImage.init(data:)`,
respectively.

Closes #119 (finally 😅).

## Changes

- Remove `Mappable`.

- Remove `JSON`, `Parse` and `Serialize` namespaces and helpers.