Implement scoped loggers from `ModuleLogger`s 🍰
On certain occasions, one needs to pass in a simple `Logger` instance
but want log events forwarded to a larger logging infrastructure.

If we're dealing with a `ModuleLogger` with multiple registered modules,
one may want to derive a simpler `Logger` that sends all events under a
single module to the upstream logger. To achieve that, `ModuleLogger`
can now create scoped loggers via a new `scopedLogger(for:)` helper.

## Changes

- Create `ModuleLogger.scopedLogger(for:)` helper.

- Create new `Log.ForwardingLogger` private helper class to forward log
events to an upstream (Module) logger.