Add single module SPM product 💧 (#232)
With the latest changes to Package.swift, importing Alicerce via SPM
resulted in `import Alicerce` no longer working, as the code is made
available in the new submodules (e.g. `AlicerceCore`,
`AlicerceLogging`, ...).

This is caused by the fact that SPM targets are the exposed modules,
and not the products (which can contain multiple targets). The solution
then becomes a matter of creating a new target that includes everything
under `Sources`, and a new product that includes just this one target.

However, as SPM performs some checks to ensure path exclusivity (to
avoid duplicate symbols), we can't simply create another target that
includes everything under `Sources`. One workaround is to trick
SPM by creating a symlink to Sources with a different name 👻
https://forums.swift.org/t/spm-shared-targets-files-use-case-whats-the-alternative/38888/4
That being said, until SPM supports more complex setups, I think we can
get away with this. Users will have to be careful to not simultaneously
import the single module product with any of the multi module products,
as that will surely cause duplicate symbols 💥.

## Changes

- Migrate Package.swift to `swift-tools-version:5.3` and fix warnings
caused by missing exclude paths in `AlicerceCore` - due to it using
`Sources` as `path`, and only `Shared` and `Utils` as sources.

- Create `SingleModuleSources` symlink to `Sources`.

- Create new `AlicerceSingleModule` product and `Alicerce` target in
`Package.swift`.

- Remove SPM shared schemes as they are made available in projects by
Xcode when an SPM module is imported... 😓