Unify Release and CI workflows 🤝
Unfortunately, `workflow_run` triggers only capture the branch that
triggered the workflow(s) they depend on, which means that we can't use
them while filtering for tag pushes (which is what we want for
triggering releases). That's why `startsWith(github.ref, 'refs/tags/')`
was never `true` when the `Release` workflow ran.

After quite a bit of experimentation and research, I came to the
conclusion that we should just keep it simple (KISS and all that 🙈)
and just unify our whole CI into a single workflow which easily
achieves the setup we want.

Hopefully third time will be the charm to get our releases working via
GitHub Actions 🍀

## Changes

- Rename `Release` workflow jobs `deploy-github` and `deploy-cocoapods`
to `release-github` and `release-cocoapods` respectively.

- Move release workflow jobs to the `CI` Workflow, configured to only
run on tag pushes (`X.Y.Z`) and with dependencies on `build-test`,
`swiftpm`, `cocoapods` and `carthage` jobs.