Make `Router` generic over its route 🚏 (#240)
Our `Router` and `RouteHandler` infrastructure was designed to work
with plain `URL`'s on its main APIs. Despite URLs being the critical
piece of information to route (e.g. in our `TrieRouter`), in certain
situations the URL doesn't contain all the information, which limits
the Routing infrastructure capabilities. One such example would be in
a server middleware router, where the HTTP method would likely be
required information to pass on to handlers to perform adequate routing.

To remove this restriction, a new `Routable` protocol was created to
represent a routable type, which only needs to provide a `url` and is
forwarded by the router to the handler upon a successful match. To make
it easier to unpack any additional information on the handlers and have
more type safety, it was made an `associatedtype` of both `Router` and
`RouteHandler`, effectively turning it into a new generic `R`.

## Changes

- Add new `Routable` protocol, add default conformance to `URL`.

- Add new `associatedtype R: Routable` to `Router` and `RouteHandler`.

- Update `Router` and `RouterHandler` APIs to receive a route of type
`R`.

- Add new `RouteHandler.eraseToAnyRouteHandler` helper.