Improve `EdgesConstrainableProxy` API ↔️
Until now, it wasn't possible to define edges constraints using a
`NSDirectionalEdgeInsets` value, requiring one to convert them to
`UIEdgeInsets`.

Additionally, the `EdgesConstrainableProxy.edges()` API only received a
single `relation` and `priority` parameters for all edges, which is
often not ideal in real scenarios. Some examples are:
 - allow a single edge constraint to "break"with a smaller `priority`
 and not others
 - allow edges to "shrink" towards the center, where the `relation`
 needs to be inverted between `leading`/`trailing` or `top`/`bottom`.

To address the above, a new overload was added to
`EdgesConstrainableProxy` that takes in a `NSDirectionalEdgeInsets`
instance, and new edge-specific relation/priority parameters were added.

## Changes

- Add a new `EdgesConstrainableProxy.edges` overload that receives a
`NSDirectionalEdgeInsets` instance instead of a `UIEdgeInsets`.

- Allow tweaking `relation` and `priority` for each edge of a
`EdgesConstrainableProxy`.

- Create helpers to convert between `UIEdgesInsets` and
`NSDirectionalEdgesInsets`.