Revise `Box`es 📦
We have had two `Box` variants for a very long time: `Box` and `VarBox`,
containing an immutable and mutable value respectively.

However, from our experience the immutable variant is not used often if
at all, which means its immutability guarantee is not something that
people want or need.

Furthermore, following the latest language developments a `Box` is a
very good candidate for a property wrapper, and more so now that they
can be defined as local properties.

## Changes

- Deprecate `VarBox` in favor of `Box`, making it wrap a mutable value.

- Make `Box` a `@propertyWrapper`, and `@dynamicMemberLookup`. The
latter is useful if we use `Box<T>` as a regular type (i.e. not a PW).