Add string log formatting witnesses and builder 👷
Until now, we relied on a custom string-based log format to convert log
items into log messages. However, it was quite limited because it only
allowed a predefined set of components/formats, evaluated the format
string for *every* log item.

By modeling formatting as witnesses wrapping a closure that can be
composed, more flexibility is achieved and items are formatted
efficiently by performing a single (composed) closure.

To improve ergonomics, a new result builder was created to allow
defining log formatting witnesses with a flexible DSL.

## Changes

- Create `LogItemFormatComponent` protocol to model first class types
for multiple composable format operations.

- Create `Log.ItemFormat` namespace, nesting:
  + Multiple `LogItemFormatComponent` implementations and static
  helpers for most common use cases.
  + New `Formatting` witnesses
  + New `Builder` result builder to build formatting witnesses.

- Update `Log.StringLogItemFormatter` to only require a `Formatting` and
be instantiated via a `FormattingBuilder`.

- Remove `LogLevelFormatter`s.

- Add `@inlinable` annotations for improved optimization possibilities.

- Deprecate `Log.Item.testItem` in favor of `Log.Item.dummy()`.

- Fix and add relevant UTs.