Fix `Log.ItemFormat` result builders in Swift 5.8 🪵
Due to recent improvements to Result Builders in Swift 5.8 / Xcode 14.3
the existing `Log.ItemFormat.Builder` implementation resulted in some
ambiguities due to it being used with two final result types:

- `Formatting<Output>`: merged formatting closure witness
- `[Formatting<Output>]`: array of formatting witnesses representing a
group to which we apply transformations, most notably adding separators
where we require the independent un-merged format group elements.

While in the previous result builder implementation the result type was
propagated to the inner builder closure, that is no longer the case in
the new version which causes the ambiguity.

Despite not being as compact, the simplest solution was to have two
different builders where each specializes in its own result type. It
causes a bit of duplication, but makes the Swift 5.8 compiler happy and
requires no changes at the point of usage of current code, while being
backwards compatible with Swift 5.7.

## Changes

- Create new `Log.ItemFormat.GroupBuilder` and use where item format
groups are required.

- Fix and add relevant UTs.

## References

- https://forums.swift.org/t/improved-result-builder-implementation-in-swift-5-8/63192
- https://github.com/apple/swift-evolution/blob/main/proposals/0289-result-builders.md#changes-from-the-first-revision