Add result builders to `MultiLogger` and `MultiTracker`s 👷 (#265)
Some of our "multiplexer" instances for logging (`Log.MultiLogger`),
analytics (`Analytics.MultiTracker`) and performance metrics
(`PerformanceMetrics.MultiTracker`) require type erasure on the child
elements to which they forward events, causing the setup to be
cumbersome and not intuitive.

By leveraging result builders we can make this a bit smoother to users
by abstracting/automating the type erasure "dance". Furthermore, it
unlocks control flow (e.g. `if/else`, `for ... in`, `if #available`)
which allows more advanced setups to be defined directly in the builder.

## Changes

- Add result builders to facilitate instantiating certain "multiplexer"
instances, most notably the ones that require type erasure:

  + `Log.MultiLogger` (requires erasing to `AnyMetadataLogDestination`)

  + `Analytics.MultiTracker` (requires erasing to `AnyAnalyticsTracker`)

  + `PerformanceMetrics.MultiTracker` (doesn't require erasing)

- Add relevant UTs.